refactor: move helper functions to tests_utils
